agentic_AI/patch_summary_agent/requirements.txt
anthropic
requests
python-dotenv
rich
jinja2
python-dateutil
